requirements.txt
openai>=1.40.0
openpyxl>=3.1.0
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
anthropic>=0.40.0          # only needed if using LLM_PROVIDER=anthropic
streamlit>=1.38.0          # web UI
